--- src/lib/index.ts
@@ -1,7 +1,0 @@
-import "./embed";
-export { useLocalMedia, useRoomConnection, VideoView } from "./react";
-export { default as fakeWebcamFrame } from "./utils/fakeWebcamFrame";
-export { default as fakeAudioStream } from "./utils/fakeAudioStream";
-export type { LocalParticipantState as LocalParticipant, WaitingParticipant, Screenshare } from "./RoomParticipant";
-export type { ChatMessage } from "./RoomConnection";
-export const sdkVersion = "__SDK_VERSION__";
