+++ src/lib/react/index.ts
@@ -1,3 +1,13 @@
 export { default as VideoView } from "./VideoView";
 export { default as useLocalMedia } from "./useLocalMedia";
 export { useRoomConnection } from "./useRoomConnection";
+export { default as LocalMedia } from "../LocalMedia";
+export { default as RoomConnection } from "../RoomConnection";
+export type {
+    RemoteParticipantState as RemoteParticipant,
+    LocalParticipantState as LocalParticipant,
+    WaitingParticipant,
+    Screenshare,
+} from "../RoomParticipant";
+export type { ChatMessage } from "../RoomConnection";
+export { sdkVersion } from "../version";
