+++ src/lib/utils/index.ts
@@ -1,0 +1,2 @@
+export { default as fakeAudioStream } from "./fakeAudioStream";
+export { default as fakeWebcamFrame } from "./fakeWebcamFrame";
