+++ src/lib/version.ts
@@ -1,0 +1,1 @@
+export const sdkVersion = "__SDK_VERSION__";
